refactor(test): move setup pocketbase to fixtures
